Fix OAuth2 by using GoogleCredentials
client_secrets.json must be included in the root directory for the app
to read from.
oauth2client/client.py was modified so that the credentials used is read
from the environment variable 'GOOGLE_APPLICATION_CREDENTIALS' instead
of AppEngine asserted credentials. This fixes a token refresh issue when
gspread tries to authorize.